move Settings to /private/ folder
